refactor: EventMenuView와 EventCreateView로 분리
